refactor: Extract offset management classes
