+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/ByteBufferEncoder.java
@@ -1,0 +1,52 @@
+package io.confluent.parallelconsumer;
+
+import java.nio.ByteBuffer;
+
+import static io.confluent.parallelconsumer.OffsetEncoding.ByteArray;
+import static io.confluent.parallelconsumer.OffsetEncoding.ByteArrayCompressed;
+
+class ByteBufferEncoder extends OffsetEncoder {
+
+    private final ByteBuffer bytesBuffer;
+
+    public ByteBufferEncoder(final int length, OffsetSimultaneousEncoder offsetSimultaneousEncoder) {
+        super(offsetSimultaneousEncoder);
+        this.bytesBuffer = ByteBuffer.allocate(1 + length);
+    }
+
+    @Override
+    protected OffsetEncoding getEncodingType() {
+        return ByteArray;
+    }
+
+    @Override
+    protected OffsetEncoding getEncodingTypeCompressed() {
+        return ByteArrayCompressed;
+    }
+
+    @Override
+    public void containsIndex(final int rangeIndex) {
+        this.bytesBuffer.put((byte) 0);
+    }
+
+    @Override
+    public void doesNotContainIndex(final int rangeIndex) {
+        this.bytesBuffer.put((byte) 1);
+    }
+
+    @Override
+    public byte[] serialise() {
+        return this.bytesBuffer.array();
+    }
+
+    @Override
+    public int getEncodedSize() {
+        return this.bytesBuffer.capacity();
+    }
+
+    @Override
+    protected byte[] getEncodedBytes() {
+        return this.bytesBuffer.array();
+    }
+
+}
